Make it harder to miss how to actually download the package
The big green button doesn't always read as a button. Instead it looks like a banner or header, which leads newbies to scroll down the page and look for the thing to click to actually download the kit. This leads to those various links to other sites (Ruby, Rails, Git...) and so people get lost, often thinking they've gotten Railsinstaller when actually they got something else.

This commit attempts to solve this by (a) making "Download it now" a link and (b) making it more clear that the list of links is *not* actually what you want to click on to download the kit.